Start splitting code into modules
