+++ lib/anime.esm.js
@@ -5,8 +5,6 @@
  * animejs.com
  */
 
-// Defaults
-
 const defaultInstanceSettings = {
   update: null,
   begin: null,
@@ -30,16 +28,30 @@ const defaultTweenSettings = {
   round: 0
 };
 
-const validTransforms = ['translateX', 'translateY', 'translateZ', 'rotate', 'rotateX', 'rotateY', 'rotateZ', 'scale', 'scaleX', 'scaleY', 'scaleZ', 'skew', 'skewX', 'skewY', 'perspective', 'matrix', 'matrix3d'];
-
-// Caching
+const validTransforms = [
+  'translateX',
+  'translateY',
+  'translateZ',
+  'rotate',
+  'rotateX',
+  'rotateY',
+  'rotateZ',
+  'scale',
+  'scaleX',
+  'scaleY',
+  'scaleZ',
+  'skew',
+  'skewX',
+  'skewY',
+  'perspective',
+  'matrix',
+  'matrix3d'
+];
 
 const cache = {
   CSS: {},
   springs: {}
 };
-
-// Utils
 
 function minMax(val, min, max) {
   return Math.min(Math.max(val, min), max);
@@ -70,8 +82,6 @@ const is = {
   col: a => (is.hex(a) || is.rgb(a) || is.hsl(a)),
   key: a => !defaultInstanceSettings.hasOwnProperty(a) && !defaultTweenSettings.hasOwnProperty(a) && a !== 'targets' && a !== 'keyframes',
 };
-
-// Easings
 
 function parseEasingParameters(string) {
   const match = /\(([^)]+)\)/.exec(string);
