+++ lib/anime.umd.js
@@ -11,8 +11,6 @@
   (global = global || self, global.anime = factory());
 }(this, (function () { 'use strict';
 
-  // Defaults
-
   const defaultInstanceSettings = {
     update: null,
     begin: null,
@@ -36,16 +34,30 @@
     round: 0
   };
 
-  const validTransforms = ['translateX', 'translateY', 'translateZ', 'rotate', 'rotateX', 'rotateY', 'rotateZ', 'scale', 'scaleX', 'scaleY', 'scaleZ', 'skew', 'skewX', 'skewY', 'perspective', 'matrix', 'matrix3d'];
-
-  // Caching
+  const validTransforms = [
+    'translateX',
+    'translateY',
+    'translateZ',
+    'rotate',
+    'rotateX',
+    'rotateY',
+    'rotateZ',
+    'scale',
+    'scaleX',
+    'scaleY',
+    'scaleZ',
+    'skew',
+    'skewX',
+    'skewY',
+    'perspective',
+    'matrix',
+    'matrix3d'
+  ];
 
   const cache = {
     CSS: {},
     springs: {}
   };
-
-  // Utils
 
   function minMax(val, min, max) {
     return Math.min(Math.max(val, min), max);
@@ -76,8 +88,6 @@
     col: a => (is.hex(a) || is.rgb(a) || is.hsl(a)),
     key: a => !defaultInstanceSettings.hasOwnProperty(a) && !defaultTweenSettings.hasOwnProperty(a) && a !== 'targets' && a !== 'keyframes',
   };
-
-  // Easings
 
   function parseEasingParameters(string) {
     const match = /\(([^)]+)\)/.exec(string);
