+++ src/cache.js
@@ -1,0 +1,6 @@
+const cache = {
+  CSS: {},
+  springs: {}
+}
+
+export default cache;
